MVP-1 Переместил константы ValueStorage
