+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/common/domain/usecase/GetModuleNameUseCase.kt
@@ -1,6 +1,6 @@
 package dev.itssho.module.qpay.module.common.domain.usecase
 
-import dev.itssho.module.hierarchy.storage.StringListKey
+import dev.itssho.module.hierarchy.storage.ValueStorage.StrList
 import dev.itssho.module.qpay.module.common.domain.storage.FullyEditableValueStorage
 import dev.itssho.module.qpay.module.common.domain.storage.UninitializedConstValueException
 
@@ -10,5 +10,5 @@ class GetModuleNameUseCase(
 
 	@Throws(UninitializedConstValueException::class)
 	operator fun invoke() : List<String> =
-		fullyEditableValueStorage.getList(StringListKey.MODULE_NAME) ?: throw UninitializedConstValueException(StringListKey.MODULE_NAME)
+		fullyEditableValueStorage.getList(StrList.MODULE_NAME) ?: throw UninitializedConstValueException(StrList.MODULE_NAME)
 }
