+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/common/domain/usecase/SetProjectPathUseCase.kt
@@ -1,6 +1,6 @@
 package dev.itssho.module.qpay.module.common.domain.usecase
 
-import dev.itssho.module.hierarchy.storage.StringListKey
+import dev.itssho.module.hierarchy.storage.ValueStorage.StrList
 import dev.itssho.module.qpay.module.common.domain.storage.FullyEditableValueStorage
 
 @Deprecated("Хз зачем это может быть нужно")
@@ -10,6 +10,6 @@ class SetProjectPathUseCase(
 
 	@Deprecated("Хз зачем это может быть нужно")
 	operator fun invoke(path: List<String>) {
-		fullyEditableValueStorage.forcePut(StringListKey.PROJECT_PATH, path)
+		fullyEditableValueStorage.forcePut(StrList.PROJECT_PATH, path)
 	}
 }
