--- ui-hierarchy/src/main/kotlin/dev/itssho/module/hierarchy/storage/StringKey.kt
@@ -1,8 +1,0 @@
-package dev.itssho.module.hierarchy.storage
-
-object StringKey {
-
-	// TODO заменить рефлексией
-	fun values(): List<String> = listOf(
-	)
-}
