+++ ui-hierarchy/src/main/kotlin/dev/itssho/module/hierarchy/storage/ValueStorageExt.kt
@@ -1,0 +1,3 @@
+package dev.itssho.module.hierarchy.storage
+
+val ValueStorage.moduleName: List<String>? get() = getList(ValueStorage.StrList.MODULE_NAME)
